[REFACTOR]: 화면 구성 디렉토리 구조 변경 - (auth), (tabs)로 분리 및 재구성
